Splitting template inheritance from Bonita base
